refactor(listener): merge listener and mod
